+++ src/App/pages/show/TicketShow.tsx
