+++ src/Components/layout/PMFLayout.tsx
@@ -1,9 +1,6 @@
 import { Layout, LayoutProps } from "react-admin";
-import { PMFAppbar } from "../appbar/PMFAppbar";
-import { PMFMenu } from "../menu/PMFMenu";
-import { Theme, useMediaQuery } from "@mui/material";
-import { Box } from "@mui/material";
-import { AppBar, Menu, Sidebar } from "react-admin";
+import { PMFAppbar } from "./appbar/PMFAppbar";
+import { PMFMenu } from "./menu/PMFMenu";
 
 export const PMFLayout = (props: LayoutProps) => (
   <Layout
